Move Cache classes to Caching namespace.
Add a top-level Cache namespace #501
